Update README
Project README has been updated with ToC, links, feature roadmap, dependencies
list, an API reference / example, and internal implementation details related to
design choices.

Additionally, the DECLARE_CRC() macro was moved to the main header because it's
useful and users shouldn't have to include an internal header to access it.