Added APIs to avoid unnecessary table inits and changed param order
This commit adds more-performant APIs that separate table initialization from
computation, allowing the user to manage it more efficiently. Unfortunately,
this inherently introduces a burden on the user to manage the table memory as
they may have multiple CRCs in use simultaneously and C limits our ability to
associate data lifetimes with the code. Adding a proper C++ API would get
around this, but is a future goal.

While adding these APIs, it seemed more logical to have the param defining the
algorithm as the first argument to all functions as it's always needed. This is
a BREAKING CHANGE to the API and required minor modifications to all test
cases.

Additionally, new documentation as been written for the functions in the
README.