Add rv-match tests for db_test and large_test
Add rv-matches to test the entirety of db_test, which required replacing the
use of nonstandard strsep() with standardized strtok(). strtok() usage isn't
ideal, but this is only a test case and it allows the verifier to actually run.

Additionally, a misc. logging statement in the CMakeLists.txt was removed and
the rv-match tests were grouped together for clarity.