Add exhaustive DECT_12 test
This commit adds a new exhaustive test case for CRC_DECT_12, which was also
added to reference.h

Additionally, a major internal change was introduced to simplify the writing of
new exhaustive test cases since the code for all of them is identical. A new
macro has been introduced that will generate the exhaustive test main.