Document verification of crc_utility functions and improve reflect() impl
Update the docs with further details on the proof status and add a new
implementation of reflect() that was produced during the proof process. This
new implementation is mathematically equivalent to the old, but up to ~260x
faster in the worst case. It's also a lot prettier to look at.