Optimize CRC for when refin && refout are true
Due to the elegance of having a single 'universal' implementation, the core
code was originally written to implement every CRC in terms of a left shift
register. However, when refin and refout are true, the underlying CRC is
actually a right shift CRC and doing this conversion requires us to
reflect every input byte as it comes in.

By adding a specialization for these CRCs, we can hoist the reflections out of
the hot loop into the setup code. This branch implements this optimization
and explores the performance implications of it.